feat: Create ExpenseEntity JPA model with fields for expense tracking, including AI categorization and receipt paths, and implement UUID generation on persist.

+++ src/main/java/com/wing/backendapiexpensespringboot/model/ExpenseEntity.java
@@ -60,7 +60,7 @@ public class ExpenseEntity {
     private UUID aiCategoryId;
 
     @Column(name = "ai_confidence")
-    private BigDecimal aiConfidence;
+    private Double aiConfidence;
 
     @Column(name = "ai_source")
     private String aiSource;
